pyproject.toml
[project]
name = "impossible-bench"
version = "0.1.0"
description = "Add your description here"
readme = "ReadMe.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.52.1",
    "click>=8.2.1",
    "huggingface-hub>=0.32.3",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "jinja2>=3.1.6",
    "openai>=1.82.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "seaborn>=0.13.2",
]
